fix: add cpp golang working
golang needs to ignore cpp file to build

+++ python/scripts/get_problem.py
@@ -18,21 +18,25 @@ from python.lc_libs import get_question_info, get_questions_by_key_word, get_que
 from python.utils import get_default_folder
 
 
-def __check_path__(problem_folder: str, problem_id: str, problem_slug: str, force: bool = False, file=None):
+def __check_path__(problem_folder: str, problem_id: str, problem_slug: str, force: bool = False,
+                   skip_language: bool = False, file=None):
     root_path = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
     dir_path = os.path.join(root_path, problem_folder, f"{problem_folder}_{problem_id}")
     if os.path.exists(dir_path):
         if not force:
             print(f"Already exists problem [{problem_id}]{problem_slug}", file=file)
             return None
+        if skip_language:
+            return dir_path
         shutil.rmtree(dir_path)
     os.mkdir(dir_path)
     return dir_path
 
 
 def process_single_algorithm_problem(problem_folder: str, problem_id: str, problem_slug: str,
-                                     problem_title: str, cookie: str, force: bool = False, file=None, languages=None):
-    dir_path = __check_path__(problem_folder, problem_id, problem_slug, force, file)
+                                     problem_title: str, cookie: str, force: bool = False, skip_language: bool = False,
+                                     file=None, languages=None):
+    dir_path = __check_path__(problem_folder, problem_id, problem_slug, force, skip_language, file)
     if not dir_path:
         return
     desc = get_question_desc(problem_slug, cookie)
@@ -52,8 +56,9 @@ def process_single_algorithm_problem(problem_folder: str, problem_id: str, probl
         return
     with open(f"{dir_path}/problem.md", "w", encoding="utf-8") as f:
         f.write(write_problem_md(problem_id, problem_title, desc))
-    with open(f"{dir_path}/testcase.py", "w", encoding="utf-8") as f:
-        f.write(write_testcase(testcases, outputs))
+    if not skip_language or not os.path.exists(f"{dir_path}/testcase.py"):
+        with open(f"{dir_path}/testcase.py", "w", encoding="utf-8") as f:
+            f.write(write_testcase(testcases, outputs))
     with open(f"{dir_path}/testcase", "w", encoding="utf-8") as f:
         f.writelines([testcase_str, "\n",
                       str(outputs).replace("None", "null")
@@ -62,15 +67,23 @@ def process_single_algorithm_problem(problem_folder: str, problem_id: str, probl
     for key, val in code_maps.items():
         match key:
             case "python3":
+                if skip_language and os.path.exists(f"{dir_path}/solution.py"):
+                    continue
                 with open(f"{dir_path}/solution.py", "w", encoding="utf-8") as f:
                     f.write(write_solution_python(val))
             case "golang":
+                if skip_language and os.path.exists(f"{dir_path}/solution.go"):
+                    continue
                 with open(f"{dir_path}/solution.go", "w", encoding="utf-8") as f:
                     f.write(write_solution_golang(val, None, problem_id))
             case "java":
+                if skip_language and os.path.exists(f"{dir_path}/Solution.java"):
+                    continue
                 with open(f"{dir_path}/Solution.java", "w", encoding="utf-8") as f:
                     f.write(write_solution_java(val, None, problem_id))
             case "cpp":
+                if skip_language and os.path.exists(f"{dir_path}/Solution.cpp"):
+                    continue
                 with open(f"{dir_path}/Solution.cpp", "w", encoding="utf-8") as f:
                     f.write(write_solution_cpp(val, None, problem_id))
             case _:
@@ -108,7 +121,8 @@ def process_single_database_problem(problem_folder: str, problem_id: str, proble
 
 def main(problem_folder: str, problem_id: Optional[str], problem_slug: Optional[str], problem_category: Optional[str],
          force: bool = False, cookie: Optional[str] = None, fetch_all: bool = False, premium_only: bool = False,
-         file: Optional[str] = None, replace_problem_id: bool = False, languages: list[str] = None):
+         file: Optional[str] = None, replace_problem_id: bool = False, skip_language: bool = False,
+         languages: list[str] = None):
     if not fetch_all:
         if not problem_id and not problem_slug:
             print("Requires at least one of problem_id or problem_slug to fetch in single mode.")
@@ -139,7 +153,7 @@ def main(problem_folder: str, problem_id: Optional[str], problem_slug: Optional[
             process_single_database_problem(problem_folder, problem_id, problem_slug, problem_title, cookie, force)
         else:
             process_single_algorithm_problem(problem_folder, problem_id, problem_slug, problem_title, cookie, force,
-                                             languages=languages)
+                                             skip_language, languages=languages)
             if replace_problem_id:
                 root_path = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
                 for lang in languages:
@@ -234,6 +248,8 @@ if __name__ == '__main__':
                         help="Debug output file, better debugging when the messages are too long", default=None)
     parser.add_argument("-change", "--change_problem_id", required=False, action="store_true",
                         help="Replace the problem id to run in each language.")
+    parser.add_argument("-sl", "--skip_language", required=False, action="store_true",
+                        help="Skip exist language files in the problem.")
     args = parser.parse_args()
 
     try:
@@ -245,5 +261,6 @@ if __name__ == '__main__':
     pf = os.getenv(constant.PROBLEM_FOLDER, get_default_folder(args.problem_category))
     langs = os.getenv(constant.LANGUAGES, "python3").split(",")
     main(pf, args.problem_id, args.problem_slug, args.problem_category,
-         args.force, cke, args.fetch_all, args.premium_only, args.debug_file, args.change_problem_id, langs)
+         args.force, cke, args.fetch_all, args.premium_only, args.debug_file, args.change_problem_id,
+         args.skip_language, langs)
     sys.exit()
